HACK: camlibs/ptp2: Work around extra file added events for first Fuji USB capture
Fuji cameras normally don't send EC_ObjectAdded messages. Except some (at
least X-T4) do send this event for the first capture after USB was plugged
in for tethered shooting. It looks like it doesn't matter how often a
session was opened/closed or if the application using libgphoto was
restarted after the camera was plugged in.

This hack assumes camera_wait_for_event running in a loop and downloading
and deleting the images as soon as they appear.